Create new Value class
